requirements.txt
slack_bolt
openai
tiktoken
fastapi
uvicorn
pillow
python-dotenv
